docs/get-started/python-quickstart/code/part-3/pyproject.toml
[project]
name = "summarize"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Cully Wakelin", email = "cully.wakelin@resonatehq.io" }
]
dependencies = [
    "resonate-sdk>=0.1.34",
    "flask>=3.0.3",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
    "app-node-1" = "summarize.app_node_1:main"
    "app-node-2" = "summarize.app_node_2"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy>=1.11.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/summarize"]

[tool.rye.scripts]
check-types = "dmypy run --timeout 60 -- src"
